Added additional validation functions
