--- UI/flask/validator.py
@@ -1,54 +1,0 @@
-"""
-This file should have independent fuctions where it validates data.
-
-
-"""
-
-
-
-def isIP(parameter):
-    """
-    Is parameter an IP?
-
-    :return:
-    """
-    return None
-
-
-def isNetwork(parameter):
-    """
-    Is parameter a network?
-
-    :return:
-    """
-    return None
-
-
-def isMask(parameter):
-    """
-    Is parameter a network?
-
-    :return:
-    """
-    return None
-
-
-def isNetwork(parameter):
-    """
-    Is parameter a network?
-
-    :return:
-    """
-    return None
-
-
-def isDomain(parameter):
-    """
-    Is parameter a domain?
-
-    :return:
-    """
-    return None
-
-
-
